chode: move GSO socket to its own struct
